backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
jsonschema==4.20.0
cadquery==2.4.0
requests>=2.28.0
python-dotenv>=1.0.0
numpy<2.0
trimesh>=4.0.0

chromadb>=0.4.15
sentence-transformers>=2.2.2
tqdm>=4.66.0
huggingface-hub>=0.20.0
Pillow>=10.0.0
